move calculation of max_mag and virtual scale into Layer.downsample
