+++ wkcuber/downsampling_utils.py
@@ -65,6 +65,27 @@ def get_next_mag(mag: Mag, scale: Optional[Tuple[float, float, float]]) -> Mag:
         )
 
 
+def calculate_virtual_scale_for_target_mag(
+    target_mag: Mag,
+) -> Tuple[float, float, float]:
+    """
+    This scale is not the actual scale of the dataset
+    The virtual scale is used for downsample_mags_anisotropic.
+    """
+    max_target_value = max(list(target_mag.to_array()))
+    scale_array = max_target_value / np.array(target_mag.to_array())
+    return cast(Tuple[float, float, float], tuple(scale_array))
+
+
+def calculate_default_max_mag(dataset_size: Tuple[int, int, int]) -> Mag:
+    # The lowest mag should have a size of ~ 100vx**3
+    max_x_y = max(dataset_size[0], dataset_size[1])
+    # highest power of 2 larger (or equal) than max_x_y divided by 100
+    return Mag(
+        max(int(math.pow(2, math.ceil(math.log(max_x_y / 100, 2)))), 4)
+    )  # at least 4
+
+
 def parse_interpolation_mode(
     interpolation_mode: str, layer_name: str
 ) -> InterpolationModes:
